+++ rust/cubesql/cubesql/e2e/tests/mysql.rs
@@ -67,10 +67,20 @@ impl MySqlIntegrationTestSuite {
         let mut table = Table::new();
         table.load_preset("||--+-++|    ++++++");
 
-        let mut header = vec![];
+        let mut header = Vec::with_capacity(res.columns_ref().len());
+        let mut description: Vec<String> = Vec::with_capacity(res.columns_ref().len());
+
         for column in res.columns_ref().into_iter() {
             header.push(Cell::new(column.name_str()));
+            description.push(format!(
+                "{} type: ({:?}:{}) flags: {:?}",
+                column.name_str(),
+                column.column_type(),
+                column.column_length(),
+                column.flags(),
+            ));
         }
+
         table.set_header(header);
 
         res.for_each(|row| {
@@ -92,7 +102,7 @@ impl MySqlIntegrationTestSuite {
         .await
         .unwrap();
 
-        table.trim_fmt()
+        description.join("\r\n").to_string() + "\r\n" + &table.trim_fmt()
     }
 
     async fn test_use(&self) -> RunResult {
@@ -165,10 +175,32 @@ impl MySqlIntegrationTestSuite {
     }
 
     fn escape_snapshot_name(&self, name: String) -> String {
-        name.to_lowercase()
+        let mut name = name
+            .to_lowercase()
             // @todo Real escape?
+            .replace("\r", "")
+            .replace("\n", "")
+            .replace("\t", "")
+            .replace(">", "")
+            .replace("<", "")
+            .replace("'", "")
+            .replace("::", "_")
+            .replace(":", "")
             .replace(" ", "_")
             .replace("*", "asterisk")
+            // shorter variant
+            .replace(",_", "_");
+
+        for _ in 0..32 {
+            name = name.replace("__", "_");
+        }
+
+        // Windows limit
+        if name.len() > 200 {
+            name.chars().into_iter().take(200).collect()
+        } else {
+            name
+        }
     }
 
     async fn assert_query(&self, conn: &mut Conn, query: String) {
@@ -201,18 +233,30 @@ impl AsyncTestSuite for MySqlIntegrationTestSuite {
         self.test_use().await?;
         self.test_prepared().await?;
         self.test_prepared_reset().await?;
-        self.test_execute_query("SELECT COUNT(*), status FROM Orders".to_string())
-            .await?;
         self.test_execute_query(
-            "SELECT COUNT(*), status, createdAt FROM Orders ORDER BY createdAt".to_string(),
+            "SELECT COUNT(*) count, status FROM Orders GROUP BY status".to_string(),
         )
         .await?;
         self.test_execute_query(
-            "SELECT COUNT(*), status, DATE_TRUNC('month', createdAt) FROM Orders ORDER BY createdAt".to_string(),
+            "SELECT COUNT(*) count, status, createdAt FROM Orders GROUP BY status, createdAt ORDER BY createdAt".to_string(),
         )
         .await?;
         self.test_execute_query(
-            "SELECT COUNT(*), status, DATE_TRUNC('quarter', createdAt) FROM Orders ORDER BY createdAt".to_string(),
+            "SELECT COUNT(*) count, status, DATE_TRUNC('month', createdAt) date FROM Orders GROUP BY status, DATE_TRUNC('month', createdAt) ORDER BY date".to_string(),
+        )
+        .await?;
+        self.test_execute_query(
+            "SELECT COUNT(*) count, status, DATE_TRUNC('quarter', createdAt) date FROM Orders GROUP BY status, DATE_TRUNC('quarter', createdAt) ORDER BY date".to_string(),
+        )
+        .await?;
+        self.test_execute_query(
+            r#"SELECT
+                CAST(true as boolean) as bool_true,
+                CAST(false as boolean) as bool_false,
+                1::int as int,
+                'str' as str
+            "#
+            .to_string(),
         )
         .await?;