+++ rust/cubesql/cubesql/src/compile/engine/information_schema/mysql/collations.rs
@@ -203,7 +203,6 @@ impl TableProvider for InfoSchemaCollationsProvider {
     async fn scan(
         &self,
         projection: &Option<Vec<usize>>,
-        _batch_size: usize,
         _filters: &[Expr],
         _limit: Option<usize>,
     ) -> Result<Arc<dyn ExecutionPlan>, DataFusionError> {