+++ rust/cubesql/cubesql/src/compile/engine/information_schema/postgres/pg_range.rs
@@ -169,7 +169,6 @@ impl TableProvider for PgCatalogRangeProvider {
     async fn scan(
         &self,
         projection: &Option<Vec<usize>>,
-        _batch_size: usize,
         _filters: &[Expr],
         _limit: Option<usize>,
     ) -> Result<Arc<dyn ExecutionPlan>, DataFusionError> {