+++ rust/cubesql/cubesql/src/config/injection.rs
@@ -1,10 +1,9 @@
 use crate::CubeError;
+
 use std::any::{type_name, TypeId};
 use std::collections::HashMap;
 use std::future::Future;
 use std::pin::Pin;
-#[allow(deprecated)]
-use std::raw::TraitObject;
 use std::sync::{Arc, Weak};
 use tokio::sync::Mutex;
 use tokio::sync::RwLock;
@@ -154,8 +153,15 @@ impl Injector {
     }
 }
 
+// TODO: Better solution
+// Rust deprecated TraitObject, because core team thinking about chaning metalayer for traits
+// It's a hacky solution to unblock upgrade of rust
+pub struct TraitObject {
+    pub data: *mut (),
+    pub vtable: *mut (),
+}
+
 pub trait DIService: Send + Sync {
-    #[allow(deprecated)]
     fn downcast_ref(
         &self,
         target: TypeId,
@@ -171,7 +177,6 @@ impl dyn DIService {
     ) -> Result<Arc<T>, CubeError> {
         unsafe {
             let obj = self.downcast_ref(TypeId::of::<T>(), type_name::<T>(), arc)?;
-            #[allow(deprecated)]
             let ptr = *(&obj as *const TraitObject as *const &T);
             Ok(Arc::from_raw(ptr))
         }
@@ -188,10 +193,11 @@ macro_rules! di_service (
                 target: core::any::TypeId,
                 type_name: &'static str,
                 arc: Arc<dyn $crate::config::injection::DIService>,
-            ) -> Result<core::raw::TraitObject, CubeError> {
+            ) -> Result<$crate::config::injection::TraitObject, CubeError> {
                 unsafe {
                     let ptr = Arc::into_raw(arc);
                     let arc = Arc::<Self>::from_raw(ptr as *const Self);
+
                     $(
                     if target == core::any::TypeId::of::<dyn $trait_ty>() {
                         let iface_arc: Arc<dyn $trait_ty> = arc;
@@ -199,10 +205,11 @@ macro_rules! di_service (
                         return Ok(std::mem::transmute(&*ptr));
                     }
                     )*
+
                     if target == core::any::TypeId::of::<$ty>() {
                         let typ_arc: Arc<$ty> = arc;
                         let ptr = Arc::into_raw(typ_arc);
-                        return Ok(core::raw::TraitObject {
+                        return Ok($crate::config::injection::TraitObject {
                             data: ptr as *const _ as *mut (),
                             vtable: std::ptr::null_mut(),
                         });