+++ rust/cubesql/cubesql/src/sql/postgres/mod.rs
@@ -1,6 +1,8 @@
 pub(crate) mod buffer;
+pub(crate) mod pg_type;
 pub(crate) mod protocol;
 pub(crate) mod service;
 pub(crate) mod shim;
 
+pub use pg_type::*;
 pub use service::*;