+++ rust/cubesql/cubesql/src/sql/types.rs
@@ -1,3 +1,4 @@
+use crate::sql::PgTypeId;
 use bitflags::bitflags;
 use msql_srv::{
     ColumnFlags as MysqlColumnFlags, ColumnType as MysqlColumnType, StatusFlags as MysqlStatusFlags,
@@ -8,6 +9,7 @@ pub enum ColumnType {
     String,
     VarStr,
     Double,
+    Boolean,
     Int8,
     Int32,
     Int64,
@@ -17,7 +19,28 @@ pub enum ColumnType {
 
 impl ColumnType {
     pub fn to_mysql(&self) -> MysqlColumnType {
-        MysqlColumnType::MYSQL_TYPE_BLOB
+        match self {
+            ColumnType::String => MysqlColumnType::MYSQL_TYPE_STRING,
+            ColumnType::VarStr => MysqlColumnType::MYSQL_TYPE_VAR_STRING,
+            ColumnType::Double => MysqlColumnType::MYSQL_TYPE_DOUBLE,
+            ColumnType::Boolean => MysqlColumnType::MYSQL_TYPE_TINY,
+            ColumnType::Int8 | ColumnType::Int32 => MysqlColumnType::MYSQL_TYPE_LONG,
+            ColumnType::Int64 => MysqlColumnType::MYSQL_TYPE_LONGLONG,
+            _ => MysqlColumnType::MYSQL_TYPE_BLOB,
+        }
+    }
+
+    pub fn to_pg_tid(&self) -> PgTypeId {
+        match self {
+            ColumnType::Blob => PgTypeId::BYTEA,
+            ColumnType::Boolean => PgTypeId::BOOL,
+            ColumnType::Int64 => PgTypeId::INT8,
+            ColumnType::Int8 => PgTypeId::INT2,
+            ColumnType::Int32 => PgTypeId::INT4,
+            ColumnType::String | ColumnType::VarStr => PgTypeId::TEXT,
+            ColumnType::Timestamp => PgTypeId::TIMESTAMP,
+            ColumnType::Double => PgTypeId::NUMERIC,
+        }
     }
 }